fix(ui): resolve conflicts
