+++ src/components/ui/badge.tsx
@@ -1,38 +1,3 @@
-import { Slot } from '@radix-ui/react-slot'
-import * as React from 'react'
-
-import { cva, type VariantProps } from 'class-variance-authority'
-
-import { cn } from '@/lib/utils'
-
-const badgeVariants = cva(
-	{
-				secondary: 'border-transparent bg-secondary text-secondary-foreground [a&]:hover:bg-secondary/90',
-				outline: 'text-foreground [a&]:hover:bg-accent [a&]:hover:text-accent-foreground',
-				destructive:
-					'border-transparent bg-destructive text-white [a&]:hover:bg-destructive/90 focus-visible:ring-destructive/20 dark:focus-visible:ring-destructive/40 dark:bg-destructive/60',
-				default: 'border-transparent bg-primary text-primary-foreground [a&]:hover:bg-primary/90',
-			},
-		},
-		defaultVariants: {
-			variant: 'default',
-		},
-	}
-)
-
-function Badge({
-	variant,
-	className,
-	asChild = false,
-	...props
-}: React.ComponentProps<'span'> & VariantProps<typeof badgeVariants> & { asChild?: boolean }) {
-	const Comp = asChild ? Slot : 'span'
-
-	return <Comp className={cn(badgeVariants({ variant }), className)} data-slot="badge" {...props} />
-}
-
-export { Badge, badgeVariants }
-
 import * as React from 'react'
 
 import { cva, type VariantProps } from 'class-variance-authority'
